requirements.txt
# FastAPI framework and server
fastapi
uvicorn[standard]

# Configuration management
pydantic-settings

# AI / LLM SDKs
together
langchain-google-genai
langchain-community
google-generativeai>=0.6.0


# Database driver
motor